Leave map benchmarks only
